pyproject.toml
[project]
name = "remindme-bot"
version = "0.1.0"
description = "Smart Reminder Telegram Bot — set reminders in natural language"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "dateparser>=1.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
